cleaned up update database dialog code
